complete history section
