+++ components/sidebar-desktop.tsx
@@ -8,10 +8,13 @@ import { ChatHistory } from '@/components/chat-history'
 import { IconPlus, IconData, IconModel, IconPromptHistory, IconTools } from '@/components/ui/icons'
 import { auth } from '@/auth'
 import { useWallet } from '@solana/wallet-adapter-react'
+import { redirect } from 'next/navigation'
 
 export function SidebarDesktop() {
   const wallet = useWallet()
-
+  const handleUrl =  () => {
+    redirect('/');
+  }
 
   return (
     <Sidebar className="peer absolute inset-y-0 z-30 hidden -translate-x-full bg-[#121212] duration-300 ease-in-out data-[state=open]:translate-x-0 lg:flex lg:w-[220px] h-full min-h-screen px-5 pt-4">
@@ -24,10 +27,7 @@ export function SidebarDesktop() {
             buttonVariants({ variant: 'outline' }),
             'h-10 w-full justify-start bg-[#171717] px-4 shadow-none transition-colors hover:bg-zinc-200/40 border-none mb-2 rounded-full ring-[3px] ring-[#1a1a1a]'
           )}
-          onClick={e => {
-            e.preventDefault();
-            window.location.reload();
-          }}
+          onClick={handleUrl}
         >
           <IconPlus className="-translate-x-2" />
           New Chat
